Refactor test, add catalog client test
