+++ axiom/connectors/ducklake/tests/CMakeLists.txt
@@ -12,7 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_executable(axiom_connectors_ducklake_test DuckLakeMetadataConfigTest.cpp DuckLakeQueryTest.cpp)
+add_executable(
+  axiom_connectors_ducklake_test
+  DuckLakeTestUtils.cpp
+  DuckLakeMetadataConfigTest.cpp
+  DuckLakeQueryTest.cpp
+  DuckLakeCatalogClientTest.cpp
+)
 
 add_test(NAME axiom_connectors_ducklake_test COMMAND axiom_connectors_ducklake_test)
